Updated OCMock, fixed build issues with Xcode 11, fixed unit tests

+++ ShareFileSDK/ShareFileSDK/Transfers/Uploaders/SFAAsyncMultiChunkFileUploader.m
@@ -160,14 +160,14 @@
 }
 
 #endif
-    unsigned long long numberOfParts;
+    NSUInteger numberOfParts;
     NSUInteger effectivePartSize = self.config.partSize;
     if (!fileLength) {
         numberOfParts = 0;
         [task taskCompletedWithError:[SFAError errorWithMessage:SFAFileReadError type:SFAErrorTypeUploadError]];
     }
     else {
-        numberOfParts = (unsigned long long)ceil((fileLength - [self.uploadSpecification.ResumeOffset unsignedLongValue]) / (effectivePartSize * 1.f));
+        numberOfParts = (NSUInteger)ceil((fileLength - [self.uploadSpecification.ResumeOffset unsignedLongValue]) / (effectivePartSize * 1.f));
     }
     if (numberOfParts > 1 && numberOfParts < self.config.numberOfThreads) {
         numberOfParts = self.config.numberOfThreads;
@@ -178,7 +178,7 @@
     unsigned long long index = self.uploadSpecification.ResumeIndex ?[self.uploadSpecification.ResumeIndex unsignedLongLongValue] : 0;
     
     NSMutableArray *fileParts = [[NSMutableArray alloc] initWithCapacity:(NSUInteger)numberOfParts];
-    for (unsigned long long i = 0; i < numberOfParts; i++) {
+    for (NSUInteger i = 0; i < numberOfParts; i++) {
         SFAFilePart *part = [[SFAFilePart alloc] init];
         part.index = index;
         part.length = effectivePartSize;
